fix(ci): 修 nullable 不能修饰多级指针；顺手清掉 E2E 测试的全部告警
- BSPDynamicHook.h: storeOld:(nullable IMP *) 是错的，nullable 只能修饰对象/
  单级指针，多级指针要用 IMP * _Nullable。这是编译错误，把 E2E job 拦下了。
- test_proxy_e2e.m: CHECK 走的是 printf，不是 NSLog，%@ 是非法转换符。
  全部改成 %s + UTF8String。
- test_proxy_e2e.m: syncGET 在 block 里直接写 out-parameter（autoreleasing），
  ARC 下可能 use-after-free；改成 __block 变量接住，信号量之后再赋出去。
- BSPProxyServer.m: NSInteger 与 NSUInteger 比较的有符号性告警。

+++ inject/probe/BSPProxyServer.m
@@ -494,7 +494,8 @@ static BOOL bsp_write_all(int fd, const void *buf, size_t len)
         }
 
         contentLength = [headers[@"content-length"] integerValue];
-        while (contentLength > 0 && (NSInteger)buf.length < NSMaxRange(hdrEnd) + contentLength) {
+        while (contentLength > 0 &&
+               (NSInteger)buf.length < (NSInteger)(NSMaxRange(hdrEnd) + (NSUInteger)contentLength)) {
             uint8_t tmp[8192];
             ssize_t n = recv(ctx.fd, tmp, sizeof(tmp), 0);
             if (n <= 0) break;
